Creation of Convex database and Clerk authorization
Created a Convex database and added a table with user information, and created functions to populate and retrieve data from the table. In addition, authorization through the Clerk was done. A webhook and event handler were created to create and edit an account and write the necessary data to the database.