Add Layouts, Navigation, Dark Mode and adding friends functionality
Functionality was created to change the theme of the application using next-themes. Two pages were created: Conversations and Friends. In the Friends component, the logic for displaying the list of friends, as well as the functionality for adding them, was created. In addition, a SideBar has been added with navigation for mobile and computer devices, where you can switch between components, set up an account and change the theme.  3 hooks were created: useConversations, useNavigation and useMutationState. useConversations is used to calculate the conversation id, useNavigation is used to display the navigation fields on the navigation menu, and useMutationState allows you to track the status of the mutation running. A functionality was also created to reject a friend request. For all the functions, the necessary functions were created in the database.